Explore data

Explore data

explore(dataset, vars = "", byvar = "", fun = c("mean", "sd"),
  top = "fun", tabfilt = "", tabsort = "", nr = NULL,
  data_filter = "", shiny = FALSE)

Arguments

dataset
Dataset to explore

vars
(Numerical) variables to summaries

byvar
Variable(s) to group data by before summarizing

fun
Functions to use for summarizing

top
The variable (type) to display at the top of the table

tabfilt
Expression used to filter the table. This should be a string (e.g., "Total > 10000")

tabsort
Expression used to sort the table (e.g., "-Total")

nr
Number of rows to display

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

shiny
Logical (TRUE, FALSE) to indicate if the function call originate inside a shiny app

Value

A list of all variables defined in the function as an object of class explore

Details

See https://radiant-rstats.github.io/docs/data/explore.html for an example in Radiant

See also

Examples

result <- explore(diamonds, "price:x")
summary(result)

#> Explore
#> Data        : diamonds 
#> Functions   : mean, sd 
#> Top         : Function 
#> 
#>  variable      mean        sd
#>     price 3,907.186 3,956.915
#>     carat     0.794     0.474
#>   clarity     0.013     0.115
#>       cut     0.034     0.180
#>     color     0.127     0.333
#>     depth    61.753     1.446
#>     table    57.465     2.241
#>         x     5.722     1.124

result <- explore(diamonds, c("price","carat"), byvar = "cut", fun = c("n_missing", "skew"))
summary(result)

#> Explore
#> Data        : diamonds 
#> Grouped by  : cut 
#> Functions   : n_missing, skew 
#> Top         : Function 
#> 
#>        cut variable n_missing  skew
#>       Fair    price         0 1.574
#>       Fair    carat         0 0.929
#>       Good    price         0 1.489
#>       Good    carat         0 1.021
#>  Very Good    price         0 1.601
#>  Very Good    carat         0 0.937
#>    Premium    price         0 1.413
#>    Premium    carat         0 0.930
#>      Ideal    price         0 1.799
#>      Ideal    carat         0 1.365

diamonds %>% explore("price", byvar = "cut", fun = c("n_obs", "n_distinct"))

#> $rng
#> [1] 2 3
#> 
#> $tab
#>         cut variable n_obs n_distinct
#> 1      Fair    price   101         98
#> 2      Good    price   275        267
#> 3 Very Good    price   677        621
#> 4   Premium    price   771        655
#> 5     Ideal    price  1176        956
#> 
#> $df_name
#> [1] "."
#> 
#> $vars
#> [1] "price"
#> 
#> $byvar
#> [1] "cut"
#> 
#> $fun
#> [1] "n_obs"      "n_distinct"
#> 
#> $top
#> [1] "fun"
#> 
#> $tabfilt
#> [1] ""
#> 
#> $tabsort
#> [1] ""
#> 
#> $nr
#> NULL
#> 
#> $data_filter
#> [1] ""
#> 
#> $shiny
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "explore" "list"