Add transformed variables to a data frame with the option to include a custom variable name extension

Add transformed variables to a data frame with the option to include a custom variable name extension

mutate_ext(.tbl, .funs, ..., .ext = "", .vars = c())

Arguments

.tbl

Data frame to add transformed variables to

.funs

Function(s) to apply (e.g., log)

...

Variables to transform

.ext

Extension to add for each variable

.vars

A list of columns generated by dplyr::vars(), or a character vector of column names, or a numeric vector of column positions.

Details

Wrapper for dplyr::mutate_at that allows custom variable name extensions

Examples

mutate_ext(mtcars, .funs = log, mpg, cyl, .ext = "_ln")
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#>                       mpg_ln   cyl_ln
#> Mazda RX4           3.044522 1.791759
#> Mazda RX4 Wag       3.044522 1.791759
#> Datsun 710          3.126761 1.386294
#> Hornet 4 Drive      3.063391 1.791759
#> Hornet Sportabout   2.928524 2.079442
#> Valiant             2.895912 1.791759
#> Duster 360          2.660260 2.079442
#> Merc 240D           3.194583 1.386294
#> Merc 230            3.126761 1.386294
#> Merc 280            2.954910 1.791759
#> Merc 280C           2.879198 1.791759
#> Merc 450SE          2.797281 2.079442
#> Merc 450SL          2.850707 2.079442
#> Merc 450SLC         2.721295 2.079442
#> Cadillac Fleetwood  2.341806 2.079442
#> Lincoln Continental 2.341806 2.079442
#> Chrysler Imperial   2.687847 2.079442
#> Fiat 128            3.478158 1.386294
#> Honda Civic         3.414443 1.386294
#> Toyota Corolla      3.523415 1.386294
#> Toyota Corona       3.068053 1.386294
#> Dodge Challenger    2.740840 2.079442
#> AMC Javelin         2.721295 2.079442
#> Camaro Z28          2.587764 2.079442
#> Pontiac Firebird    2.954910 2.079442
#> Fiat X1-9           3.306887 1.386294
#> Porsche 914-2       3.258097 1.386294
#> Lotus Europa        3.414443 1.386294
#> Ford Pantera L      2.760010 2.079442
#> Ferrari Dino        2.980619 1.791759
#> Maserati Bora       2.708050 2.079442
#> Volvo 142E          3.063391 1.386294
mutate_ext(mtcars, .funs = log, .ext = "_ln")
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#>                       mpg_ln   cyl_ln  disp_ln    hp_ln  drat_ln     wt_ln
#> Mazda RX4           3.044522 1.791759 5.075174 4.700480 1.360977 0.9631743
#> Mazda RX4 Wag       3.044522 1.791759 5.075174 4.700480 1.360977 1.0560527
#> Datsun 710          3.126761 1.386294 4.682131 4.532599 1.348073 0.8415672
#> Hornet 4 Drive      3.063391 1.791759 5.552960 4.700480 1.124930 1.1678274
#> Hornet Sportabout   2.928524 2.079442 5.886104 5.164786 1.147402 1.2354715
#> Valiant             2.895912 1.791759 5.416100 4.653960 1.015231 1.2412686
#> Duster 360          2.660260 2.079442 5.886104 5.501258 1.166271 1.2725656
#> Merc 240D           3.194583 1.386294 4.988390 4.127134 1.305626 1.1600209
#> Merc 230            3.126761 1.386294 4.947340 4.553877 1.366092 1.1474025
#> Merc 280            2.954910 1.791759 5.121580 4.812184 1.366092 1.2354715
#> Merc 280C           2.879198 1.791759 5.121580 4.812184 1.366092 1.2354715
#> Merc 450SE          2.797281 2.079442 5.619676 5.192957 1.121678 1.4036430
#> Merc 450SL          2.850707 2.079442 5.619676 5.192957 1.121678 1.3164082
#> Merc 450SLC         2.721295 2.079442 5.619676 5.192957 1.121678 1.3297240
#> Cadillac Fleetwood  2.341806 2.079442 6.156979 5.323010 1.075002 1.6582281
#> Lincoln Continental 2.341806 2.079442 6.131226 5.370638 1.098612 1.6908336
#> Chrysler Imperial   2.687847 2.079442 6.086775 5.438079 1.172482 1.6761615
#> Fiat 128            3.478158 1.386294 4.365643 4.189655 1.406097 0.7884574
#> Honda Civic         3.414443 1.386294 4.326778 3.951244 1.595339 0.4793350
#> Toyota Corolla      3.523415 1.386294 4.264087 4.174387 1.439835 0.6070445
#> Toyota Corona       3.068053 1.386294 4.788325 4.574711 1.308333 0.9021918
#> Dodge Challenger    2.740840 2.079442 5.762051 5.010635 1.015231 1.2584610
#> AMC Javelin         2.721295 2.079442 5.717028 5.010635 1.147402 1.2340169
#> Camaro Z28          2.587764 2.079442 5.857933 5.501258 1.316408 1.3454724
#> Pontiac Firebird    2.954910 2.079442 5.991465 5.164786 1.124930 1.3467736
#> Fiat X1-9           3.306887 1.386294 4.369448 4.189655 1.406097 0.6601073
#> Porsche 914-2       3.258097 1.386294 4.789989 4.510860 1.488400 0.7608058
#> Lotus Europa        3.414443 1.386294 4.554929 4.727388 1.327075 0.4140944
#> Ford Pantera L      2.760010 2.079442 5.860786 5.575949 1.439835 1.1537316
#> Ferrari Dino        2.980619 1.791759 4.976734 5.164786 1.286474 1.0188473
#> Maserati Bora       2.708050 2.079442 5.707110 5.814131 1.264127 1.2725656
#> Volvo 142E          3.063391 1.386294 4.795791 4.691348 1.413423 1.0224509
#>                      qsec_ln vs_ln am_ln  gear_ln   carb_ln
#> Mazda RX4           2.800933  -Inf     0 1.386294 1.3862944
#> Mazda RX4 Wag       2.834389  -Inf     0 1.386294 1.3862944
#> Datsun 710          2.923699     0     0 1.386294 0.0000000
#> Hornet 4 Drive      2.967333     0  -Inf 1.098612 0.0000000
#> Hornet Sportabout   2.834389  -Inf  -Inf 1.098612 0.6931472
#> Valiant             3.006672     0  -Inf 1.098612 0.0000000
#> Duster 360          2.762538  -Inf  -Inf 1.098612 1.3862944
#> Merc 240D           2.995732     0  -Inf 1.386294 0.6931472
#> Merc 230            3.131137     0  -Inf 1.386294 0.6931472
#> Merc 280            2.906901     0  -Inf 1.386294 1.3862944
#> Merc 280C           2.939162     0  -Inf 1.386294 1.3862944
#> Merc 450SE          2.856470  -Inf  -Inf 1.098612 1.0986123
#> Merc 450SL          2.867899  -Inf  -Inf 1.098612 1.0986123
#> Merc 450SLC         2.890372  -Inf  -Inf 1.098612 1.0986123
#> Cadillac Fleetwood  2.889260  -Inf  -Inf 1.098612 1.3862944
#> Lincoln Continental 2.880321  -Inf  -Inf 1.098612 1.3862944
#> Chrysler Imperial   2.857619  -Inf  -Inf 1.098612 1.3862944
#> Fiat 128            2.968875     0     0 1.386294 0.0000000
#> Honda Civic         2.918851     0     0 1.386294 0.6931472
#> Toyota Corolla      2.990720     0     0 1.386294 0.0000000
#> Toyota Corona       2.996232     0  -Inf 1.098612 0.0000000
#> Dodge Challenger    2.825537  -Inf  -Inf 1.098612 0.6931472
#> AMC Javelin         2.850707  -Inf  -Inf 1.098612 0.6931472
#> Camaro Z28          2.735017  -Inf  -Inf 1.098612 1.3862944
#> Pontiac Firebird    2.836150  -Inf  -Inf 1.098612 0.6931472
#> Fiat X1-9           2.939162     0     0 1.386294 0.0000000
#> Porsche 914-2       2.815409  -Inf     0 1.609438 0.6931472
#> Lotus Europa        2.827314     0     0 1.609438 0.6931472
#> Ford Pantera L      2.674149  -Inf     0 1.609438 1.3862944
#> Ferrari Dino        2.740840  -Inf     0 1.609438 1.7917595
#> Maserati Bora       2.681022  -Inf     0 1.609438 2.0794415
#> Volvo 142E          2.923162     0     0 1.386294 0.6931472
mutate_ext(mtcars, .funs = log)
#>                          mpg      cyl     disp       hp     drat        wt
#> Mazda RX4           3.044522 1.791759 5.075174 4.700480 1.360977 0.9631743
#> Mazda RX4 Wag       3.044522 1.791759 5.075174 4.700480 1.360977 1.0560527
#> Datsun 710          3.126761 1.386294 4.682131 4.532599 1.348073 0.8415672
#> Hornet 4 Drive      3.063391 1.791759 5.552960 4.700480 1.124930 1.1678274
#> Hornet Sportabout   2.928524 2.079442 5.886104 5.164786 1.147402 1.2354715
#> Valiant             2.895912 1.791759 5.416100 4.653960 1.015231 1.2412686
#> Duster 360          2.660260 2.079442 5.886104 5.501258 1.166271 1.2725656
#> Merc 240D           3.194583 1.386294 4.988390 4.127134 1.305626 1.1600209
#> Merc 230            3.126761 1.386294 4.947340 4.553877 1.366092 1.1474025
#> Merc 280            2.954910 1.791759 5.121580 4.812184 1.366092 1.2354715
#> Merc 280C           2.879198 1.791759 5.121580 4.812184 1.366092 1.2354715
#> Merc 450SE          2.797281 2.079442 5.619676 5.192957 1.121678 1.4036430
#> Merc 450SL          2.850707 2.079442 5.619676 5.192957 1.121678 1.3164082
#> Merc 450SLC         2.721295 2.079442 5.619676 5.192957 1.121678 1.3297240
#> Cadillac Fleetwood  2.341806 2.079442 6.156979 5.323010 1.075002 1.6582281
#> Lincoln Continental 2.341806 2.079442 6.131226 5.370638 1.098612 1.6908336
#> Chrysler Imperial   2.687847 2.079442 6.086775 5.438079 1.172482 1.6761615
#> Fiat 128            3.478158 1.386294 4.365643 4.189655 1.406097 0.7884574
#> Honda Civic         3.414443 1.386294 4.326778 3.951244 1.595339 0.4793350
#> Toyota Corolla      3.523415 1.386294 4.264087 4.174387 1.439835 0.6070445
#> Toyota Corona       3.068053 1.386294 4.788325 4.574711 1.308333 0.9021918
#> Dodge Challenger    2.740840 2.079442 5.762051 5.010635 1.015231 1.2584610
#> AMC Javelin         2.721295 2.079442 5.717028 5.010635 1.147402 1.2340169
#> Camaro Z28          2.587764 2.079442 5.857933 5.501258 1.316408 1.3454724
#> Pontiac Firebird    2.954910 2.079442 5.991465 5.164786 1.124930 1.3467736
#> Fiat X1-9           3.306887 1.386294 4.369448 4.189655 1.406097 0.6601073
#> Porsche 914-2       3.258097 1.386294 4.789989 4.510860 1.488400 0.7608058
#> Lotus Europa        3.414443 1.386294 4.554929 4.727388 1.327075 0.4140944
#> Ford Pantera L      2.760010 2.079442 5.860786 5.575949 1.439835 1.1537316
#> Ferrari Dino        2.980619 1.791759 4.976734 5.164786 1.286474 1.0188473
#> Maserati Bora       2.708050 2.079442 5.707110 5.814131 1.264127 1.2725656
#> Volvo 142E          3.063391 1.386294 4.795791 4.691348 1.413423 1.0224509
#>                         qsec   vs   am     gear      carb
#> Mazda RX4           2.800933 -Inf    0 1.386294 1.3862944
#> Mazda RX4 Wag       2.834389 -Inf    0 1.386294 1.3862944
#> Datsun 710          2.923699    0    0 1.386294 0.0000000
#> Hornet 4 Drive      2.967333    0 -Inf 1.098612 0.0000000
#> Hornet Sportabout   2.834389 -Inf -Inf 1.098612 0.6931472
#> Valiant             3.006672    0 -Inf 1.098612 0.0000000
#> Duster 360          2.762538 -Inf -Inf 1.098612 1.3862944
#> Merc 240D           2.995732    0 -Inf 1.386294 0.6931472
#> Merc 230            3.131137    0 -Inf 1.386294 0.6931472
#> Merc 280            2.906901    0 -Inf 1.386294 1.3862944
#> Merc 280C           2.939162    0 -Inf 1.386294 1.3862944
#> Merc 450SE          2.856470 -Inf -Inf 1.098612 1.0986123
#> Merc 450SL          2.867899 -Inf -Inf 1.098612 1.0986123
#> Merc 450SLC         2.890372 -Inf -Inf 1.098612 1.0986123
#> Cadillac Fleetwood  2.889260 -Inf -Inf 1.098612 1.3862944
#> Lincoln Continental 2.880321 -Inf -Inf 1.098612 1.3862944
#> Chrysler Imperial   2.857619 -Inf -Inf 1.098612 1.3862944
#> Fiat 128            2.968875    0    0 1.386294 0.0000000
#> Honda Civic         2.918851    0    0 1.386294 0.6931472
#> Toyota Corolla      2.990720    0    0 1.386294 0.0000000
#> Toyota Corona       2.996232    0 -Inf 1.098612 0.0000000
#> Dodge Challenger    2.825537 -Inf -Inf 1.098612 0.6931472
#> AMC Javelin         2.850707 -Inf -Inf 1.098612 0.6931472
#> Camaro Z28          2.735017 -Inf -Inf 1.098612 1.3862944
#> Pontiac Firebird    2.836150 -Inf -Inf 1.098612 0.6931472
#> Fiat X1-9           2.939162    0    0 1.386294 0.0000000
#> Porsche 914-2       2.815409 -Inf    0 1.609438 0.6931472
#> Lotus Europa        2.827314    0    0 1.609438 0.6931472
#> Ford Pantera L      2.674149 -Inf    0 1.609438 1.3862944
#> Ferrari Dino        2.740840 -Inf    0 1.609438 1.7917595
#> Maserati Bora       2.681022 -Inf    0 1.609438 2.0794415
#> Volvo 142E          2.923162    0    0 1.386294 0.6931472
mutate_ext(mtcars, .funs = log, .ext = "_ln", .vars = vars(mpg, cyl))
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#>                       mpg_ln   cyl_ln
#> Mazda RX4           3.044522 1.791759
#> Mazda RX4 Wag       3.044522 1.791759
#> Datsun 710          3.126761 1.386294
#> Hornet 4 Drive      3.063391 1.791759
#> Hornet Sportabout   2.928524 2.079442
#> Valiant             2.895912 1.791759
#> Duster 360          2.660260 2.079442
#> Merc 240D           3.194583 1.386294
#> Merc 230            3.126761 1.386294
#> Merc 280            2.954910 1.791759
#> Merc 280C           2.879198 1.791759
#> Merc 450SE          2.797281 2.079442
#> Merc 450SL          2.850707 2.079442
#> Merc 450SLC         2.721295 2.079442
#> Cadillac Fleetwood  2.341806 2.079442
#> Lincoln Continental 2.341806 2.079442
#> Chrysler Imperial   2.687847 2.079442
#> Fiat 128            3.478158 1.386294
#> Honda Civic         3.414443 1.386294
#> Toyota Corolla      3.523415 1.386294
#> Toyota Corona       3.068053 1.386294
#> Dodge Challenger    2.740840 2.079442
#> AMC Javelin         2.721295 2.079442
#> Camaro Z28          2.587764 2.079442
#> Pontiac Firebird    2.954910 2.079442
#> Fiat X1-9           3.306887 1.386294
#> Porsche 914-2       3.258097 1.386294
#> Lotus Europa        3.414443 1.386294
#> Ford Pantera L      2.760010 2.079442
#> Ferrari Dino        2.980619 1.791759
#> Maserati Bora       2.708050 2.079442
#> Volvo 142E          3.063391 1.386294